fix(control-plane): Address PR review comments
Two findings, both fixed.

The date the conversion is anchored on came from
`toLocaleDateString('en-CA')`, on the assumption that this locale yields
YYYY-MM-DD. That is an ICU data detail, not something the locale
guarantees. Were the output to shift, the template string built from it
would be unparseable, `naive` would be NaN, every later arithmetic step
would stay NaN, and the failure would surface as a RangeError out of
.toISOString() -- a 500 on an endpoint that only wanted to render a date.
The date now comes from the same `formatToParts` read the offset
measurement already used, as numbers into Date.UTC(), which has no such
failure mode. Verified across seven zones spanning UTC-3:30 to UTC+14,
including the half- and quarter-hour offsets: all resolve to exactly
03:00 local. The 2026-03-29 Europe/Zurich spring-forward still maps a
non-existent 02:30 forward to 03:30, matching zoneinfo and the JDK.

The function existed as three byte-identical copies. The two API copies
now import from _utils/timezone.js. The worker's copy had no caller and
was deleted rather than wired up: Terraform deploys the worker as a raw
single file (`content = file(...)`, tofu/control-plane/main.tf:51) with no
bundler, so an import there would not resolve -- that copy could never
have shared the module, which is presumably how it drifted into being
dead in the first place. The rationale for measuring the offset from the
full local date rather than hh:mm -- the bug this PR fixes -- moved into
the shared module with it.
