+++ control-plane/functions/api/info.js
@@ -7,6 +7,7 @@
  */
 import { fetchWithTimeout } from './_utils/fetch-with-timeout.js';
 import { ALL_SPINUP_WORKFLOWS, ALL_TEARDOWN_WORKFLOWS } from './_utils/workflow-selection.js';
+import { timeInTimezoneToUTC } from './_utils/timezone.js';
 
 // D1 Helper Functions
 async function getConfig(db, key, defaultValue = null) {
@@ -16,74 +17,6 @@ async function getConfig(db, key, defaultValue = null) {
   } catch {
     return defaultValue;
   }
-}
-
-/**
- * Convert a time in a specific timezone to UTC Date
- * @param {string} timeStr - Time in HH:MM format
- * @param {string} timezone - IANA timezone (e.g., 'Europe/Zurich')
- * @param {Date} baseDate - Base date to use (defaults to today)
- * @returns {Date} - Date object representing the time in UTC
- */
-function timeInTimezoneToUTC(timeStr, timezone, baseDate = new Date()) {
-  const [hours, minutes] = timeStr.split(':').map(Number);
-
-  // The local calendar date in the target zone, e.g. "2026-08-27".
-  const dateStr = baseDate.toLocaleDateString('en-CA', { timeZone: timezone });
-
-  // The wanted wall-clock moment, read as if it were UTC. Not the answer —
-  // it is off by exactly the zone's offset, which the passes below measure.
-  const naive = new Date(
-    `${dateStr}T${String(hours).padStart(2, '0')}:${String(minutes).padStart(2, '0')}:00Z`
-  ).getTime();
-
-  // Measure the offset by formatting an instant in the zone and reading the
-  // FULL local date back, not just hours and minutes.
-  //
-  // Reading only hh:mm is what made "next teardown" show tomorrow: for a
-  // late local time the probe lands on the next day in the zone (22:00Z is
-  // 00:00 in Europe/Zurich), so 22:00 versus 00:00 measured +1320 minutes
-  // instead of -120. Folding that modulo a day fixes Zurich but stays
-  // ambiguous near ±12h, where UTC+14 and UTC+10 give the same remainder.
-  const offsetMsAt = (instant) => {
-    const parts = new Intl.DateTimeFormat('en-CA', {
-      timeZone: timezone,
-      hour12: false,
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-      second: '2-digit',
-    }).formatToParts(instant);
-
-    const f = {};
-    for (const p of parts) {
-      if (p.type !== 'literal') f[p.type] = p.value;
-    }
-
-    // hour comes back as "24" rather than "00" at midnight in some runtimes.
-    const localAsUTC = Date.UTC(
-      Number(f.year),
-      Number(f.month) - 1,
-      Number(f.day),
-      Number(f.hour) % 24,
-      Number(f.minute),
-      Number(f.second)
-    );
-    return localAsUTC - instant.getTime();
-  };
-
-  // Two passes, because the offset has to be the one in force at the ANSWER,
-  // not at the probe. They differ across a daylight-saving transition: on
-  // 2026-03-29 in Europe/Zurich, 01:30Z is already 03:30 local, so a single
-  // pass measures +2h and returns 00:30 local for a requested 01:30.
-  //
-  // The second pass re-measures at the candidate instant and converges. A
-  // clock time that does not exist — 02:30 on a spring-forward day — maps
-  // forward to 03:30, which is what zoneinfo and the JDK also do.
-  const firstPass = naive - offsetMsAt(new Date(naive));
-  return new Date(naive - offsetMsAt(new Date(firstPass)));
 }
 
 export async function onRequestGet(context) {
